third_party/blink/web_tests/fast/dom/SelectorAPI/only-shadow-host-in-shadow-tree-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
crbug.com/359854: test for traversing elements in shadow tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.querySelector('body /deep/ #target').id is "target"
PASS successfullyParsed is true

TEST COMPLETE
